better names for build & task files
